Allow serialization_lib to use an external flatbuffers target
Before this change the only way to build the serialization_lib was to
have the flatbuffers lib pre-built before calling cmake or to allow
serialization_lib to build its own flatbuffers library.  This commit
make sure that you can build serialization_lib as part of the cmake
build that builds flatbuffers (not in advance) by checking that
flatbuffers is in the TARGET list.  In that case it is simply linked
by the serialization_lib library.

Signed-off-by: Davide Grohmann <davide.grohmann@arm.com>
Change-Id: I5a9d5f263ccfb36273977d6e6a2c8083598b0afb